mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.57.1",
    "mcp>=1.10.1",
    "python-dotenv>=1.1.1",
]
